+++ spec/lib/aixm/document_spec.rb
@@ -513,7 +513,7 @@ describe AIXM::Document do
               <codeType>RADIO</codeType>
               <noSeq>1</noSeq>
             </AhaUid>
-            <txtAddress>123.35</txtAddress>
+            <txtAddress>123.35 mhz</txtAddress>
             <txtRmk>A/A (callsign PUJAUT)</txtRmk>
           </Aha>
           <!-- Airspace: [D] POLYGON AIRSPACE -->
@@ -1331,7 +1331,7 @@ describe AIXM::Document do
               <codeType>RADIO</codeType>
               <noSeq>1</noSeq>
             </AhaUid>
-            <txtAddress>123.35</txtAddress>
+            <txtAddress>123.35 mhz</txtAddress>
             <txtRmk>A/A (callsign PUJAUT)</txtRmk>
           </Aha>
           <!-- Airspace: [D] POLYGON AIRSPACE -->
@@ -2197,7 +2197,7 @@ describe AIXM::Document do
               <codeType>RADIO</codeType>
               <noSeq>1</noSeq>
             </AhaUid>
-            <txtAddress>123.35</txtAddress>
+            <txtAddress>123.35 mhz</txtAddress>
             <txtRmk>A/A (callsign PUJAUT)</txtRmk>
           </Aha>
           <!-- Airspace: [D] POLYGON AIRSPACE -->