+++ spec/lib/aixm/feature/address_spec.rb
@@ -5,6 +5,12 @@ describe AIXM::Feature::Address do
     AIXM::Factory.address
   end
 
+  describe :initialize do
+    it "fails for type :radio_frequency if address is no AIXM::F" do
+      _{ AIXM.address(type: :radio_frequency, address: "foobar") }.must_raise ArgumentError
+    end
+  end
+
   describe :type= do
     it "fails on invalid values" do
       _{ subject.type = :foobar }.must_raise ArgumentError
@@ -14,6 +20,26 @@ describe AIXM::Feature::Address do
     it "looks up valid values" do
       _(subject.tap { _1.type = :phone }.type).must_equal :phone
       _(subject.tap { _1.type = :RADIO }.type).must_equal :radio_frequency
+    end
+  end
+
+  describe :address= do
+    it "fails on invalid values" do
+      _{ subject.address = nil }.must_raise ArgumentError
+    end
+
+    it "stringifies valid values for types other than :radio_frequency" do
+      _(AIXM.address(type: :other, address: 123).address).must_equal '123'
+    end
+  end
+
+  describe :address do
+    it "returns AIXM::F for type :radio_frequency" do
+      _(subject.tap { _1.type = :radio_frequency; _1.address = AIXM.f(123.45, :mhz) }.address).must_equal AIXM.f(123.45, :mhz)
+    end
+
+    it "returns String for all other types" do
+      _(subject.tap { _1.type = :other; _1.address = 'foobar' }.address).must_equal 'foobar'
     end
   end
 
@@ -31,7 +57,7 @@ describe AIXM::Feature::Address do
             <codeType>RADIO</codeType>
             <noSeq>1</noSeq>
           </XxxUid>
-          <txtAddress>123.35</txtAddress>
+          <txtAddress>123.35 mhz</txtAddress>
           <txtRmk>A/A (callsign PUJAUT)</txtRmk>
         </Xxx>
       END