+++ spec/lib/aixm/feature/airport_spec.rb
@@ -474,7 +474,7 @@ describe AIXM::Feature::Airport do
             <codeType>RADIO</codeType>
             <noSeq>1</noSeq>
           </AhaUid>
-          <txtAddress>123.35</txtAddress>
+          <txtAddress>123.35 mhz</txtAddress>
           <txtRmk>A/A (callsign PUJAUT)</txtRmk>
         </Aha>
         <!-- Address: URL for LFNT -->